ac-mlops-pipeline/clv_mlops_pipeline/requirements.txt
mlflow==2.11.3
numpy>=1.23.0
pandas==1.5.3
scikit-learn>=1.1.1
matplotlib>=3.5.2
pillow>=10.0.1
Jinja2==3.0.3
pyspark~=3.3.0
pytz~=2022.2.1
pytest>=7.1.2
pyyaml==6.0.2